requirements.txt
# Для заданий 1-2
boto3>=1.34.0
aiobotocore>=2.5.0
botocore>=1.29.0

# Для задания 3
pandas>=2.0.0
numpy>=1.24.0
watchdog>=3.0.0  # для мониторинга файлов
openpyxl>=3.1.0  # для чтения Excel
pyarrow>=14.0.0  # для чтения Parquet
xlrd>=2.0.0      # для чтения старых Excel файлов

# Общие
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Утилиты
tqdm>=4.65.0     # для прогресс-баров (опционально)